Finished factory implementation
Has user input the alphabet once, then generates encoder and decoder
functions that use the alphabet.  The default base-62 alphabet is used
if none is provided.  The advantage with having factory functions is
that we only have to do the 'tableization' of the alphabet *once*, so
subsequent decoding is very efficient.  Also, the encode/decode
functions have a more natural API (string / int).